+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 10.8
 - HTML add "Check All" checkbox after Periods in AddAbsences.php
 - Truncate Assignment title to 36 chars in AnomalousGrades.php
 - SQL resources table: Add PUBLISHED_PROFILES & PUBLISHED_GRADE_LEVELS columns in Update.fnc.php, rosariosis.sql & rosariosis_mysql.sql
+- Add Resource Visibility options in Resources.php
 
 Changes in 10.7.1
 -----------------